m_Constant 멤버 변수 double로 수정

+++ ImageProcessing/CConstantDlg.h
@@ -22,5 +22,6 @@ protected:
 
 	DECLARE_MESSAGE_MAP()
 public:
-	int m_Constant;
+//	int m_Constant;
+	double m_Constant;
 };
